webapp/backend/pyproject.toml
[project]
name = "sbai-backend"
version = "1.0.0"
description = "SB선보 P&ID 도면 AI 변환 시스템 - Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "ezdxf>=1.3.0",
    "matplotlib>=3.9.0",
    "numpy>=1.26.0",
    "openpyxl>=3.1.0",
    "PyMuPDF>=1.24.0",
    "openai>=1.60.0",
    "anthropic>=0.40.0",
    "google-genai>=1.0.0",
    "aiosqlite>=0.19.0",
    "pillow>=10.0.0",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"
